requirements.txt
torch==2.2.0
pyyaml==6.0.1
numpy==1.26.4
